pyproject.toml
[project]
name = "medical-ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "google-genai>=1.47.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
]
